# docs/openapi.yaml
# File: openapi.yaml
# Purpose: Source of truth for the REST contract and generated TypeScript schema types.
# Components: requests, resources, pagination, dashboard, audit events, and RFC 9457 errors.
openapi: 3.1.0
info:
  title: Integration Operations Console API
  version: 1.0.0
  description: Deterministic connector reliability simulator. All providers and accounts are fictional.
servers:
  - url: http://localhost:8080
tags:
  - name: Discovery
  - name: Accounts
  - name: Jobs
  - name: Operations
paths:
  /api/v1/connectors:
    get:
      tags: [Discovery]
      operationId: listConnectors
      responses:
        "200":
          description: Available fake connectors
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ConnectorDescriptor" }
  /api/v1/accounts:
    get:
      tags: [Accounts]
      operationId: listAccounts
      responses:
        "200":
          description: Linked fake accounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ConnectorAccount" }
    post:
      tags: [Accounts]
      operationId: createAccount
      parameters:
        - $ref: "#/components/parameters/DemoRole"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAccountRequest" }
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountMutationResult" }
        "200":
          description: Existing account returned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountMutationResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
  /api/v1/jobs:
    get:
      tags: [Jobs]
      operationId: listJobs
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        "200":
          description: Keyset-paginated jobs
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobPage" }
        "400": { $ref: "#/components/responses/Problem" }
    post:
      tags: [Jobs]
      operationId: submitJob
      parameters:
        - $ref: "#/components/parameters/DemoRole"
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, pattern: "^[A-Za-z0-9._:-]{8,100}$" }
        - in: header
          name: X-Demo-Actor
          schema: { type: string, maxLength: 100 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubmitJobRequest" }
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobMutationResult" }
        "200":
          description: Existing idempotent job returned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobMutationResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/jobs/run-next:
    post:
      tags: [Operations]
      operationId: runNextJob
      parameters:
        - $ref: "#/components/parameters/DemoRole"
      responses:
        "200":
          description: Scheduler result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunNextResult" }
        "403": { $ref: "#/components/responses/Problem" }
  /api/v1/jobs/{id}:
    get:
      tags: [Jobs]
      operationId: getJob
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncJob" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/jobs/{id}/run:
    post:
      tags: [Operations]
      operationId: runJob
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/DemoRole"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncJob" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/jobs/{id}/cancel:
    post:
      tags: [Operations]
      operationId: cancelJob
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/DemoRole"
      responses:
        "200":
          description: Cancelled job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncJob" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/jobs/{id}/replay:
    post:
      tags: [Operations]
      operationId: replayJob
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/DemoRole"
      responses:
        "200":
          description: Requeued job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncJob" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/jobs/{id}/audit:
    get:
      tags: [Jobs]
      operationId: listJobAudit
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Chronological append-only audit events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuditEvent" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/dashboard:
    get:
      tags: [Operations]
      operationId: getDashboard
      responses:
        "200":
          description: Operational summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dashboard" }
components:
  parameters:
    JobId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    DemoRole:
      in: header
      name: X-Demo-Role
      required: true
      description: Demonstration-only role boundary; not production authentication.
      schema: { type: string, enum: [viewer, operator] }
  responses:
    Problem:
      description: RFC 9457 problem detail
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    JobStatus:
      type: string
      enum: [QUEUED, RUNNING, RETRY_WAIT, SUCCEEDED, DEAD_LETTERED, CANCELLED]
    FailurePlan:
      type: string
      enum: [NONE, RATE_LIMIT_ONCE, TRANSIENT_TWICE, PERMANENT]
    ConnectorDescriptor:
      type: object
      required: [id, displayName]
      properties:
        id: { type: string }
        displayName: { type: string }
    ConnectorAccount:
      type: object
      required: [id, connectorId, displayName, externalReference, createdAt]
      properties:
        id: { type: string, format: uuid }
        connectorId: { type: string }
        displayName: { type: string }
        externalReference: { type: string }
        createdAt: { type: string, format: date-time }
    CreateAccountRequest:
      type: object
      additionalProperties: false
      required: [connectorId, displayName, externalReference]
      properties:
        connectorId: { type: string, pattern: "^[a-z0-9-]{2,40}$" }
        displayName: { type: string, maxLength: 100 }
        externalReference: { type: string, pattern: "^[A-Za-z0-9._-]{2,120}$" }
    SubmitJobRequest:
      type: object
      additionalProperties: false
      required: [accountId, failurePlan, maxAttempts]
      properties:
        accountId: { type: string, format: uuid }
        failurePlan: { $ref: "#/components/schemas/FailurePlan" }
        maxAttempts: { type: integer, minimum: 1, maximum: 10 }
    SyncJob:
      type: object
      required:
        - id
        - accountId
        - status
        - failurePlan
        - attempt
        - maxAttempts
        - availableAt
        - correlationId
        - idempotencyKey
        - recordsProcessed
        - createdAt
        - updatedAt
      properties:
        id: { type: string, format: uuid }
        accountId: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/JobStatus" }
        failurePlan: { $ref: "#/components/schemas/FailurePlan" }
        attempt: { type: integer, minimum: 0 }
        maxAttempts: { type: integer, minimum: 1 }
        pageCursor: { type: [string, "null"] }
        availableAt: { type: string, format: date-time }
        correlationId: { type: string }
        idempotencyKey: { type: string }
        lastError: { type: [string, "null"] }
        recordsProcessed: { type: integer, minimum: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AuditEvent:
      type: object
      required: [id, jobId, eventType, actor, detail, occurredAt]
      properties:
        id: { type: integer, format: int64 }
        jobId: { type: string, format: uuid }
        eventType: { type: string }
        actor: { type: string }
        detail: { type: string }
        occurredAt: { type: string, format: date-time }
    AccountMutationResult:
      type: object
      required: [resource, created]
      properties:
        resource: { $ref: "#/components/schemas/ConnectorAccount" }
        created: { type: boolean }
    JobMutationResult:
      type: object
      required: [resource, created]
      properties:
        resource: { $ref: "#/components/schemas/SyncJob" }
        created: { type: boolean }
    JobPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/SyncJob" }
        nextCursor: { type: [string, "null"] }
    RunNextResult:
      type: object
      required: [workAvailable]
      properties:
        job:
          oneOf:
            - { $ref: "#/components/schemas/SyncJob" }
            - { type: "null" }
        workAvailable: { type: boolean }
    Dashboard:
      type: object
      required: [statusCounts, recentJobs]
      properties:
        statusCounts:
          type: object
          additionalProperties: { type: integer, format: int64 }
        recentJobs:
          type: array
          items: { $ref: "#/components/schemas/SyncJob" }
    Problem:
      type: object
      required: [status, title, detail]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }